requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

sentence-transformers
faiss-cpu

groq
streamlit